'Add Printer' form added to 'Admin Settings'
It doesn't add printer yet this is just so I have something up in case
you guys wanted to look at it.  You can add extra fields but there is a
bug where the extra fields end up in the background of the image every
time you add a new one.  Will try to fix it as soon as I can.