Refactored printer dialog classes
"You got database queries in my dialog classes!"
"YOU got UI in my database queries!"

... Not any more! ZCorpDialog, SolidscapeDialog, and ObjetDialog have
been refactored so don't directly query the database or move files
around! Instead, it has UtilController preform those duties, ain't that
neato?

I added method called submitBuildInfoToDB. It is called in the
submitBtnActionPerformed methods of the dialog classes to take the
information entered by the user and enter it into the database. There
are still a few things dependent on the printer in use, which is not
preferable. It's pretty clear where the method is printer dependent
though, as there are switch statements determining which printer is in
use. Hopefully it won't be too difficult to change when the dynamic DB
is in place. Most of the method isn't printer dependent.
Also things about determining price have been removed BUT since there
are still columns about price in the DB, I have put "0.00" in as a
placeholder. This can be seen when you look at the reports page.
Obviously this isn't the best solution, but my knowledge of creating a
database is frankly limited, and I didn't want to mess up the data base.

I also added a method to the FileUtils method in ObjectLabSoftware, to
fix a problem I was having with moving files. This is further explained
in the documentation of the method.

I also took it upon myself to start writing more in depth documentation
for the methods I create at least. I think that we need to make an
effort to make the code more readable and easier to understand,
considering we aren't going to be the last class working on this. So far
all I have done is write up more in depth documentation for the headers
of methods. I've included the purpose of the method, where it's called,
and a small description of each parameter. I want to continue doing this
for more methods. Please let me know what you think or if there's a
better way to do it.

Sorry for the novel. I felt it was best to thouroughly explain what I
did and the reasoning behind it.

Happy coding,

David "The Nicholas Cage of Coding" Prince