Refactored to remove Printer Main classes
As it says in the title, the following classes have been removed from
the project:
- ObjetMain
- PrinterOpenBuilds
- SolidscapeMain
- ZCorpMain

Two methods have been added to UtilController to make this possible:

1. startPrintBuild this class does exactly what it says, it starts the
printBuild class. It does exactly what pressing the "Enter a new build"
button did in the printer Main classes. This is called from TomSoftMain

2. returnHome: because the main classes are now gone, there needed to be
a new way to return to the homescreen, so this method was created. It is
called whenever either the PrinterBuild window or a Dialog window is
closed.

I have tested the submitting a build process multiple times with both
the Solidscape and Zcorp printers and haven't had any hitches. All the
reports appear and all the data is in the DB.

The next step is to further polish the changes I have made in the last
three commits. They work (or so I think), it's time to make them pretty!

Happy coding,

David "The Class Destroyer" Prince