Moved returnHome and edited revertBuild
I moved the returnHome method, previously in UtilController into
PrinterBuild and each of the Dialog methods.

I also further edited revertBuild and removed that disgusting switch
statement that was previously there. I was able to do this by using a
previously unused method in SQLMethods, searchPrintersByBuildName. I
edited the method so it would use a string for the printer as a
parameter and used that string to complete the SQL method.