Dynamic reports
So I finished adding the code to make the reports show 100% dynamically
based on a SQL statement. So the headers are selected, the data loads,
you can filter based on the headers(sql columns), and to top it off it
exports what is in the grid to excel with the headers!

Emily helped with this. We did the first commit on her computer.