MasterReportExport
I added a button to the report page that will export all data for all
printers. The report saves in one excel file with each printer being the
name of the sheet and the data being inside of that sheet. The file is
saved in the export folder and has the name "MasterReport" for the
moment.