Small changes to SQLMethods according to Diagram
I made a few small changes to the SQLMethods class according to the
Database Diagram. I was going to go further but came across some bugs
regarding adding a class or submitting into the system as a student. I
ran into problems with the select statements using "current" although it
does not come up in the newest diagram on GitHub. We need to figure out
what we want to do with  "current" because it shows up often in our code
even though. It seems as though someone deleted it from the diagram but
not in the code. Unless this happened in another branch, then we're all
good.