#23 on waffle- Created button for Archiving Folders
I created a method to archive folders at the end of the semester. It
will tell you whether or not the archive was successful or if it failed
and saves it to a zipped folder in a Sync folder in your local C drive.
Still needs to be made dynamic.
-Miguel worked with me