Bug fixes that should have been given a branch
So I forgot to create a different branch to fix all of these, but I
suppose that's not too much of a problem right now considering I'm the
only on working on this? Anyways in the future I'll be trying to fix
bugs on separate branches.

The first bug fix was fixing a UI error in which the invalid student ID
prompt would stay up after you logged into the administrator page.

Fixed a bug so the browse function for the build page would open to the
toPrint page for the selected printer by default.

Changed UI so the Help selection in the menu bar would be to the right
of anything else.